Vf working partially.  Not quite right

+++ dvipdfm/tfm.c
@@ -1,4 +1,4 @@
-/*  $Header: /home/mwicks/Projects/Gaspra-projects/cvs2darcs/Repository-for-sourceforge/dvipdfm/tfm.c,v 1.7 1998/12/08 05:33:35 mwicks Exp $
+/*  $Header: /home/mwicks/Projects/Gaspra-projects/cvs2darcs/Repository-for-sourceforge/dvipdfm/tfm.c,v 1.8 1998/12/09 04:04:30 mwicks Exp $
 
     This is dvipdf, a DVI to PDF translator.
     Copyright (C) 1998  by Mark A. Wicks
@@ -281,8 +281,7 @@ void tfm_close_all(void)
 }
 
 /* tfm_get_width returns the width of the font
-   as a fraction of the design size */
-
+   as a (double) fraction of the design size */
 double tfm_get_width (int font_id, UNSIGNED_PAIR ch)
 {
   UNSIGNED_QUAD charinfo;
